Renamed files, components and updated Contacts Table
