+++ app/javascript/src/components/Dashboard/Contacts/Pane/Form.jsx
@@ -1,13 +1,13 @@
 import React, { useState } from "react";
 
-import { Formik, Form } from "formik";
+import { Formik, Form as FormikForm } from "formik";
 import { Check } from "neetoicons";
 import { Button, Pane, Toastr } from "neetoui";
 import { Input, Select } from "neetoui/formik";
 
 import { CONTACTS_VALIDATION, ROLES_VALUES } from "../constants";
 
-const ContactForm = ({ onClose, contact, isEdit }) => {
+const Form = ({ onClose, contact, isEdit }) => {
   const [submitted, setSubmitted] = useState(false);
 
   const handleSubmit = () => {
@@ -24,7 +24,7 @@ const ContactForm = ({ onClose, contact, isEdit }) => {
       onSubmit={handleSubmit}
     >
       {({ isSubmitting }) => (
-        <Form className="w-full">
+        <FormikForm className="w-full">
           <Pane.Body className="space-y-6">
             <div className="flex w-full justify-between space-x-4">
               <Input
@@ -82,10 +82,10 @@ const ContactForm = ({ onClose, contact, isEdit }) => {
               onClick={onClose}
             />
           </Pane.Footer>
-        </Form>
+        </FormikForm>
       )}
     </Formik>
   );
 };
 
-export default ContactForm;
+export default Form;
